Remove unused modal and move ModalEdit
